Fixed tabbar visibility when returning to HomeTabView and address other reviewer comments
