+++ TravelScheduler/TravelScheduler/Models/StationModel.swift
@@ -8,9 +8,6 @@
 import Foundation
 
 struct Station: Identifiable, Hashable {
-    var id = UUID()
-    var title: String
+    let id = UUID()
+    let title: String
 }
-
-
-
