+++ TravelScheduler/TravelScheduler/ViewModels/CarrierViewModel.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class CarriersViewModel: ObservableObject {
+final class CarriersViewModel: ObservableObject {
     @Published var carriers: [CarrierModel] = []
     
     init() {
@@ -15,11 +15,7 @@ class CarriersViewModel: ObservableObject {
     }
     
     func loadCarriers() {
-        carriers = [
-            CarrierModel(name: "РЖД", logo: "RZD", email: "info@rzd.ru", phone: "+74952222222"),
-            CarrierModel(name: "ФГК", logo: "FGK", email: "info@fgk.ru", phone: "+74953333333"),
-            CarrierModel(name: "Урал Логистика", logo: "URAL", email: "info@ural.ru", phone: "+74954444444")
-        ]
+        carriers = MockData.carriers
     }
     
     func carrier(for name: String) -> CarrierModel? {
