+++ TravelScheduler/TravelScheduler/ViewModels/StationsViewModel.swift
@@ -7,43 +7,16 @@
 
 import Foundation
 
-class StationsViewModel: ObservableObject {
-    @Published var cities: [City] = [
-        City(title: "Москва", stations: [
-            Station(title: "Курский Вокзал"),
-            Station(title: "Казанский Вокзал"),
-            Station(title: "Киевский Вокзал"),
-            Station(title: "Ленинградский Вокзал")
-        ]),
-        City(title: "Санкт-Петербург", stations: [
-            Station(title: "Московский Вокзал"),
-            Station(title: "Витебский Вокзал"),
-            Station(title: "Ладожский Вокзал"),
-            Station(title: "Балтийский Вокзал")
-        ]),
-        City(title: "Сочи", stations: [
-            Station(title: "Вокзал Сочи"),
-            Station(title: "Вокзал Адлер"),
-            Station(title: "Вокзал Лазаревская")
-        ]),
-        City(title: "Краснодар", stations: [
-            Station(title: "Краснодар I"),
-            Station(title: "Краснодар II")
-        ]),
-        City(title: "Горный Воздух", stations: [
-            Station(title: "Горно-Алтайск"),
-            Station(title: "Бийск"),
-            Station(title: "Белокуриха")
-        ]),
-        City(title: "Омск", stations: [
-            Station(title: "Омск-Пассажирский"),
-            Station(title: "Омск-Западный")
-        ]),
-        City(title: "Казань", stations: [
-            Station(title: "Казанский Вокзал"),
-            Station(title: "Вокзал Казань-Южный")
-        ])
-    ]
+final class StationsViewModel: ObservableObject {
+    @Published var cities: [City] = []
+    
+    init() {
+        loadCities()
+    }
+    
+    func loadCities() {
+        cities = MockData.cities
+    }
     
     func city(for station: String) -> String {
         for city in cities {
